+++ src/Manufactures.Domain/FabricConstructions/ReadModels/FabricConstructionReadModel.cs
@@ -8,9 +8,9 @@ namespace Manufactures.Domain.FabricConstructions.ReadModels
         public string ConstructionNumber { get; internal set; }
         public string MaterialType { get; internal set; }
         public string WovenType { get; internal set; }
-        public double AmountOfWarp { get; internal set; }
-        public double AmountOfWeft { get; internal set; }
-        public double Width { get; internal set; }
+        public int AmountOfWarp { get; internal set; }
+        public int AmountOfWeft { get; internal set; }
+        public int Width { get; internal set; }
         public string WarpType { get; internal set; }
         public string WeftType { get; internal set; }
         public double ReedSpace { get; internal set; }
